Started test for corpus preprocessor
